+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Version 0.2.16 (next release)
 * Made LinOp for multiplication on the right.
 * TODO break up DCPAttr so only use atom DCP function.
 * Added boolean and integer variables.
+* Made geo_mean elementwise.
 
 Version 0.2.15
 ------------